MAINT: Move suspected segfaulting test into its own file.
It looks like the test often segfault on the 26th test of
test_qt_viewer, which appear to be this one  – I can't reproduce so ran
locally with ``-v``, and move the test to a separate file to see if it
segfaults there.

If that's the case I can tag the test and run it in a separate matrix
item.
